# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Crypto Gateway API",
    "version": "0.1.0",
    "summary": "Non-custodial inbound stablecoin payments: merchant payment intents and quotes, operator evidence and oversight.",
    "description": "Every amount is a decimal string of integer units; no money ever passes through a JSON number. Merchant writes require an Idempotency-Key header scoped to the authenticated merchant, route and key. Errors share one envelope. Timestamps are RFC 3339.",
    "license": { "name": "Apache-2.0", "identifier": "Apache-2.0" }
  },
  "servers": [{ "url": "https://gateway.example.com", "description": "TLS terminates in front of the API" }],
...
  "components": {
    "securitySchemes": {
      "merchantKey": { "type": "http", "scheme": "bearer", "description": "A merchant API key of 32 to 256 characters. Stored hashed; never recoverable from the database." },
      "operatorKey": { "type": "http", "scheme": "bearer", "description": "An operator API key of 32 to 256 characters with a least-privilege subset of ingest, risk_ingest, read and admin. A risk_ingest key is additionally bound in the database to its KYT provider. Never a merchant key." }
    },
    "parameters": {
...
      "CheckoutToken": { "name": "checkout_token", "in": "path", "required": true, "schema": { "type": "string", "pattern": "^[0-9a-f]{64}$" } },
...
    },
...
    "schemas": {
      "IntegerString": { "type": "string", "pattern": "^[0-9]+$", "description": "An integer as a decimal string. Never a JSON number." },
      "Timestamp": { "type": "string", "format": "date-time", "description": "RFC 3339, UTC." },
      "Error": {
        "type": "object", "required": ["error"],
        "properties": { "error": { "type": "object", "required": ["code", "message"], "properties": {
          "code": { "type": "string", "enum": ["authentication_failed", "invalid_request", "invalid_idempotency_key", "payment_intent_not_found", "idempotency_conflict", "payment_intent_reference_conflict", "payment_intent_not_quotable", "quote_unavailable", "rail_stopped", "missing_scope", "no_price_policy", "no_open_rail_stop", "price_not_agreed", "storage_unavailable", "internal_error"] },
          "message": { "type": "string" } } } }
      },
      "FiatAmount": { "type": "object", "required": ["currency", "minor_units"], "properties": { "currency": { "type": "string", "pattern": "^[A-Z]{3}$" }, "minor_units": { "$ref": "#/components/schemas/IntegerString" } } },
...
      "CheckoutView": { "type": "object", "required": ["status", "merchant_name", "fiat_amount", "asset", "collector_address", "amount_raw", "amount", "received", "expires_at", "late_payment_until"],
        "properties": {
          "status": { "type": "string", "enum": ["paid", "needs_review", "underpaid", "cancelled", "confirming", "expired", "waiting"], "description": "paid means settled and the merchant told; confirming means a matching transfer is on the chain and not yet final or settled; needs_review means money arrived that a person must look at." },
          "merchant_name": { "type": "string" }, "description": { "type": ["string", "null"] },
          "fiat_amount": { "$ref": "#/components/schemas/FiatAmount" },
          "asset": { "$ref": "#/components/schemas/IssuedQuote/properties/asset" },
          "collector_address": { "type": "string" },
          "amount_raw": { "$ref": "#/components/schemas/IntegerString" },
          "amount": { "type": "string", "description": "amount_raw in whole tokens, exact." },
          "received": { "type": "string", "description": "Credited to this attempt so far, in whole tokens." },
          "expires_at": { "$ref": "#/components/schemas/Timestamp" }, "late_payment_until": { "$ref": "#/components/schemas/Timestamp" },
          "transaction_hash": { "type": ["string", "null"] }, "explorer_url": { "type": ["string", "null"] }
        } },
...
    },
    "responses": {
...
      "StorageUnavailable": { "description": "The database could not answer. Nothing was decided.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
    }
  },
  "paths": {
...
    "/v1/checkout/{checkout_token}": {
      "get": { "tags": ["checkout"], "summary": "The hosted payment page's state", "description": "Public: the per-attempt token from the quote is the only key, and it opens a read of that attempt only. Responses are not cacheable and send no referrer. A malformed and an unknown token answer alike.", "security": [],
        "parameters": [{ "$ref": "#/components/parameters/CheckoutToken" }],
        "responses": { "200": { "description": "The buyer-facing state.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CheckoutView" } } } }, "404": { "description": "checkout_not_found.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }, "503": { "$ref": "#/components/responses/StorageUnavailable" } } }
    },
    "/v1/checkout/{checkout_token}/qr.svg": {
      "get": { "tags": ["checkout"], "summary": "The payment address as a QR code", "description": "Encodes the address only; the buyer enters the exact amount shown beside it, because wallets differ in which payment URI fields they honour.", "security": [],
        "parameters": [{ "$ref": "#/components/parameters/CheckoutToken" }],
        "responses": { "200": { "description": "SVG image.", "content": { "image/svg+xml": { "schema": { "type": "string" } } } }, "404": { "description": "checkout_not_found.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }, "503": { "$ref": "#/components/responses/StorageUnavailable" } } }
    },
    "/checkout/{checkout_token}": {
      "get": { "tags": ["checkout"], "summary": "The hosted payment page", "description": "A static page under a strict content security policy (own script, style and image only). It reads its state from /v1/checkout/{checkout_token}.", "security": [],
        "parameters": [{ "$ref": "#/components/parameters/CheckoutToken" }],
        "responses": { "200": { "description": "HTML.", "content": { "text/html": { "schema": { "type": "string" } } } } } }
    },
...
  }
}